requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
Flask-JWT-Extended==4.7.1
pymongo==4.13.2
dnspython==2.7.0
python-dotenv==1.1.1
